+++ CHANGELOG.md
@@ -1,23 +1,39 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [4.4.0](https://github.com/imaginary-cloud/CameraManager/tree/4.4.0) - 2019-03-12
+## [5.0.0](https://github.com/imaginary-cloud/CameraManager/tree/5.0.0) - 2020-04-01
+
 ### Changed
+
+- Syntax update for Swift 5.2 (issue #189)
+
+### Fixed
+
+- Fix Swift Package Manager support (issue #215)
+
+## [4.4.0](https://github.com/imaginary-cloud/CameraManager/tree/4.4.0) - 2019-03-12
+
+### Changed
+
 - Change callbacks to the swifty way of reporting success and errors (pull request #185)
 
 ## [4.3.1](https://github.com/imaginary-cloud/CameraManager/tree/4.3.1) - 2019-03-08
+
 ### Fixed
 
 - Swift Support (pull request #173 and #168)
 - Example App layout supports iPhone X (pull request #173)
 
 ## [4.3.0](https://github.com/imaginary-cloud/CameraManager/tree/4.3.0) - 2018-08-17
+
 ### Added
 
 - Add video stabilization mode (pull request #160)
 
 ## [4.2.3](https://github.com/imaginary-cloud/CameraManager/tree/4.2.3) - 2018-07-27
+
 ### Fixed
 
 - Deinit cause crash when `cameraDevice = .front` (#157)
@@ -27,6 +43,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add note to documentation when a property has a default value
 
 ## [4.2.2](https://github.com/imaginary-cloud/CameraManager/tree/4.2.2) - 2018-07-26
+
 ### Added
 
 - Add `CONTRIBUTING.md`
@@ -35,8 +52,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - When recording is finished `videoURL` is always nil (#158)
 
-
 ## [4.2.1](https://github.com/imaginary-cloud/CameraManager/tree/4.2.1) - 2018-06-20
+
 ### Added
 
 - Add `CODE_OF_CONDUCT.md`
@@ -46,6 +63,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fix flash for video (#152)
 
 ## [4.2.0](https://github.com/imaginary-cloud/CameraManager/tree/4.2.0) - 2018-06-19
+
 ### Added
 
 - Add properties for setting custom album names for image and video recordings (pull request #147)
@@ -58,11 +76,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Memory leak in Example App
 
 ## [4.1.1](https://github.com/imaginary-cloud/CameraManager/tree/4.1.1) - 2018-05-16
+
 ### Fixed
 
 - License
 
 ## [4.1.0](https://github.com/imaginary-cloud/CameraManager/tree/4.1.0) - 2018-05-15
+
 ### Added
 
 - Add EXIF metadata for images and video - including location (#131)
@@ -71,10 +91,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - Fix Crash (pull request #141)
-- Fix image orientation when taking pictures with device rotation lock active 
+- Fix image orientation when taking pictures with device rotation lock active
 
 ## [4.0.2](https://github.com/imaginary-cloud/CameraManager/tree/4.0.2) - 2018-05-07
-### Fixed 
+
+### Fixed
 
 - Using CoreMotion to get real device orientation, fixing no sound when switching camera (pull request #134)
 - Fixed take picture on iPad with front camera(landscape, UpsideDown… (pull request #139)
@@ -88,15 +109,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Repeat crash CameraManager.swift line 430 in release (issue #138)
 
 ## [4.0.1](https://github.com/imaginary-cloud/CameraManager/tree/4.0.1) - 2017-11-18
+
 ### Added
 
 - Add @discardableResult modifiers to addPreviewLayerToView (pull request #132)
 
-### Fixed 
+### Fixed
 
 - Fix shouldEnableTapToFocus function (pull request #133)
 
 ## [4.0.0](https://github.com/imaginary-cloud/CameraManager/tree/4.0.0) - 2017-10-22
+
 ### Changed
 
 - Syntax update for Swift 4.0 (pull request #125)
@@ -106,12 +129,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add gesture recognizers on the main thread (pull request #123)
 
 ## [3.2.0](https://github.com/imaginary-cloud/CameraManager/tree/3.2.0) - 2017-07-03
+
 ### Added
 
 - Add location data to videos (pull request #110)
 - Optional location permissions (pull request #110)
 
 ## [3.1.4](https://github.com/imaginary-cloud/CameraManager/tree/3.1.4) - 2017-06-14
+
 ### Added
 
 - Add properties `focusMode` and `exposureMode` (pull request #106)
@@ -124,33 +149,39 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Getting same video URL, when simultaneously recording video (issue #108)
 
 ## [3.1.3](https://github.com/imaginary-cloud/CameraManager/tree/3.1.3) - 2017-05-15
+
 ### Added
 
 - Add two new properties `shouldEnableTapToFocus` and `shouldEnablePinchToZoom` (pull request #106)
 
 ## [3.1.2](https://github.com/imaginary-cloud/CameraManager/tree/3.1.2) - 2017-05-02
+
 ### Changed
 
 - New option to flip image took by front camera (pull request #104)
 - Fixes possible hang after requesting permission (pull request #98)
 
 ## [3.1.1](https://github.com/imaginary-cloud/CameraManager/tree/3.1.1) - 2017-03-15
+
 ### Changed
 
 - Refactor to avoid implicit unwrapped optionals (pull request #94)
 
 ## [3.1.0](https://github.com/imaginary-cloud/CameraManager/tree/3.1.0) - 2017-02-11
+
 ### Added
 
 - Flip animation and tap to focus (pull request #72)
 - Icons and splash image to example
 
 ## [3.0.0](https://github.com/imaginary-cloud/CameraManager/tree/3.0.0) - 2016-09-16
+
 ### Changed
 
 - Syntax update for Swift 3.0.
 
 ## [2.2.4](https://github.com/imaginary-cloud/CameraManager/tree/2.2.4) - 2016-07-06
+
 ### Added
 
 - Add error checking.
@@ -160,6 +191,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fixes completion typos and suggests renamed functions.
 
 ## [2.2.3](https://github.com/imaginary-cloud/CameraManager/tree/2.2.3) - 2016-05-12
+
 ### Changed
 
 - Fixed zoom in StillImage Mode.
@@ -167,16 +199,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Minor refactoring
 
 ## [2.2.2](https://github.com/imaginary-cloud/CameraManager/tree/2.2.2) - 2016-03-07
+
 ### Added
 
 - `CHANGELOG.md` file.
 
 ## [2.2.1](https://github.com/imaginary-cloud/CameraManager/tree/2.2.1) - 2016-03-02
+
 ### Added
 
 - Initial support for the Swift Package Manager.
 
 ## [2.2.0](https://github.com/imaginary-cloud/CameraManager/tree/2.2.0) - 2016-02-19
+
 ### Added
 
 - Zoom support.
@@ -186,6 +221,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed spelling of `embeddingView`.
 
 ## [2.1.3](https://github.com/imaginary-cloud/CameraManager/tree/2.1.3) - 2016-01-08
+
 ### Changed
 
 - No sound in video with more than 10 seconds fixed.
@@ -193,6 +229,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed `NewCameraOutputMode` not passed during init.
 
 ## [2.1.2](https://github.com/imaginary-cloud/CameraManager/tree/2.1.2) - 2015-12-24
+
 ### Added
 
 - Property `cameraIsReady`.
@@ -200,36 +237,43 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Completion block `addPreviewLayerToView`.
 
 ## [2.1.1](https://github.com/imaginary-cloud/CameraManager/tree/2.1.1) - 2015-12-11
+
 ### Added
 
 - Ability to disable responding to device orientation changes.
 
 ## [2.1.0](https://github.com/imaginary-cloud/CameraManager/tree/2.1) - 2015-11-20
+
 ### Added
 
 - Properties `recordedDuration` and `recordedFileSize`.
 
 ## [2.0.2](https://github.com/imaginary-cloud/CameraManager/tree/2.0.2) - 2015-11-17
+
 ### Fixed
 
 - iOS 9.0.1 bug.
 
 ## [2.0.1](https://github.com/imaginary-cloud/CameraManager/tree/2.0.1) - 2015-09-17
+
 ### Changed
 
 - Syntax updates.
 
 ## [2.0.0](https://github.com/imaginary-cloud/CameraManager/tree/2.0.0) - 2015-07-30
+
 ### Changed
 
 - Syntax update for Swift 2.0.
 
 ## [1.0.14](https://github.com/imaginary-cloud/CameraManager/tree/1.0.14) - 2015-07-17
+
 ### Changed
 
 - Small fixes.
 
 ## [1.0.13](https://github.com/imaginary-cloud/CameraManager/tree/1.0.13) - 2015-05-12
+
 ### Added
 
 - Support for installing via Carthage.
@@ -241,6 +285,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Syntax update for Swift 1.2.
 
 ## [1.0.12](https://github.com/imaginary-cloud/CameraManager/tree/1.0.12) - 2015-03-23
+
 ### Added
 
 - Incremental flash mode.
@@ -254,6 +299,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `README.md` update.
 
 ## [1.0.11](https://github.com/imaginary-cloud/CameraManager/tree/1.0.11) - 2015-03-20
+
 ### Added
 
 - Property `showAccessPermissionPopupAutomatically`, to determine if you want the user to be asked about camera permissions automatically or manually.
@@ -261,6 +307,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Error handling in capture completion blocks.
 
 ## [1.0.10](https://github.com/imaginary-cloud/CameraManager/tree/1.0.10) - 2015-03-19
+
 ### Added
 
 - Camera state returned when adding the preview layer.
@@ -270,11 +317,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `README.md` update.
 
 ## [1.0.9](https://github.com/imaginary-cloud/CameraManager/tree/1.0.9) - 2015-03-10
+
 ### Changed
 
 - CameraManager class made public.
 
 ## [1.0.8](https://github.com/imaginary-cloud/CameraManager/tree/1.0.8) - 2015-02-24
+
 ### Fixed
 
 - Wrong orientation when camera preview starts in landscape mode.
@@ -286,6 +335,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Bug which prevented from recording audio.
 
 ## [1.0.7](https://github.com/imaginary-cloud/CameraManager/tree/1.0.7) - 2014-10-30
+
 ### Added
 
 - Version compatible with XCode 6.1.
@@ -297,6 +347,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Swift syntax updates to resolve compile errors.
 
 ## [1.0.6](https://github.com/imaginary-cloud/CameraManager/tree/1.0.6) - 2014-10-28
+
 ### Added
 
 - Check for valid capture session.
@@ -306,6 +357,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fixed video orientation change.
 
 ## [1.0.5](https://github.com/imaginary-cloud/CameraManager/tree/1.0.5) - 2014-10-22
+
 ### Changed
 
 - Enhanced Camera lifecyle.
@@ -313,6 +365,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Orientation observers only added if needed.
 
 ## [1.0.4](https://github.com/imaginary-cloud/CameraManager/tree/1.0.4) - 2014-10-16
+
 ### Added
 
 - Restart session.
@@ -322,9 +375,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `README.md` update.
 
 ## [1.0.3](https://github.com/imaginary-cloud/CameraManager/tree/1.0.3) - 2014-10-15
+
 ### Added
 
--  Property `writeFilesToPhoneLibrary` to conditionally write to user library.
+- Property `writeFilesToPhoneLibrary` to conditionally write to user library.
 
 ### Changed
 
@@ -333,16 +387,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `README.md` update.
 
 ## [1.0.2](https://github.com/imaginary-cloud/CameraManager/tree/1.0.2) - 2014-10-15
+
 ### Added
 
 - `CameraManager.podspec` file.
 
 ## [1.0.1](https://github.com/imaginary-cloud/CameraManager/tree/1.0.1) - 2014-10-15
+
 ### Changed
 
 - Optional initializer for `addPreviewLayerToView`.
 
 ## [1.0.0](https://github.com/imaginary-cloud/CameraManager/tree/1.0.0) - 2014-10-15
+
 ### Added
 
 - Front and back camera selection.